+++ src/components/mainMenu/OfflineSubmenu.tsx
@@ -1,7 +1,7 @@
 import { JSX } from 'react';
 import { Dropdown } from 'react-bootstrap';
 import { BiWifiOff } from 'react-icons/bi';
-import { FaEraser } from 'react-icons/fa';
+import { FaEraser, FaRegCheckCircle, FaRegCircle } from 'react-icons/fa';
 import { useMessages } from '../../l10nInjector.js';
 import { CacheMode } from '../../types/common.js';
 import { Checkbox } from '../Checkbox.js';
@@ -28,7 +28,6 @@ export function OfflineSubmenu({
       <Dropdown.Item
         as="button"
         eventKey="caching-active-toggle"
-        active={cachingActive}
         disabled={cacheMode === 'cacheOnly'}
       >
         <Checkbox value={cachingActive} /> {m?.offline.cachingActive}
@@ -42,39 +41,28 @@ export function OfflineSubmenu({
 
       <Dropdown.Header>{m?.offline.dataSource}</Dropdown.Header>
 
-      <Dropdown.Item
-        as="button"
-        eventKey="cacheMode-networkOnly"
-        active={cacheMode === 'networkOnly'}
-      >
-        <Checkbox value={cacheMode === 'networkOnly'} />{' '}
+      <Dropdown.Item as="button" eventKey="cacheMode-networkOnly">
+        {cacheMode === 'networkOnly' ? <FaRegCheckCircle /> : <FaRegCircle />}{' '}
         {m?.offline.networkOnly}
       </Dropdown.Item>
 
-      <Dropdown.Item
-        as="button"
-        eventKey="cacheMode-networkFirst"
-        active={cacheMode === 'networkFirst'}
-      >
-        <Checkbox value={cacheMode === 'networkFirst'} />{' '}
+      <Dropdown.Item as="button" eventKey="cacheMode-networkFirst">
+        {cacheMode === 'networkFirst' ? <FaRegCheckCircle /> : <FaRegCircle />}{' '}
         {m?.offline.networkFirst}
       </Dropdown.Item>
 
-      <Dropdown.Item
-        as="button"
-        eventKey="cacheMode-cacheFirst"
-        active={cacheMode === 'cacheFirst'}
-      >
-        <Checkbox value={cacheMode === 'cacheFirst'} /> {m?.offline.cacheFirst}
+      <Dropdown.Item as="button" eventKey="cacheMode-cacheFirst">
+        {cacheMode === 'cacheFirst' ? <FaRegCheckCircle /> : <FaRegCircle />}{' '}
+        {m?.offline.cacheFirst}
       </Dropdown.Item>
 
       <Dropdown.Item
         as="button"
         eventKey="cacheMode-cacheOnly"
-        active={cacheMode === 'cacheOnly'}
         disabled={!cacheExists}
       >
-        <Checkbox value={cacheMode === 'cacheOnly'} /> {m?.offline.cacheOnly}
+        {cacheMode === 'cacheOnly' ? <FaRegCheckCircle /> : <FaRegCircle />}{' '}
+        {m?.offline.cacheOnly}
       </Dropdown.Item>
     </>
   );
